requirements.txt
flask
qrcode
pillow
pyinstaller